// frontend/src/core/model/RawDelegateInfo.ts
export type RawDelegateInfo = {
    address: string;
    blockHeight: number;
    consensus: boolean;
    ddkVersion: string;
    feeReward: number;
    forgedBlocks: number;
    location: string;
    missedBlocks: number;
    peers: number;
    uptime: number;
    votes: number;
    username: string;
};